refactored classes
